third_party/WebKit/LayoutTests/fast/forms/select/listbox-height-with-before.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<p>Selecting menu list options using up/down key should fire onchange event</p>

<style>
select:before {
    content: "";
}
</style>
<select id="listbox" multiple>
  <option>option 1</option>
  <option>option 2</option>
  <option>option 3</option>
  <option>option 4</option>
</select>

<script>
shouldBeGreaterThan('document.getElementById("listbox").offsetHeight', '2');
</script>
</body>
</html>
