third_party/blink/web_tests/platform/generic/fast/forms/select/listbox-height-with-before-expected.txt

PASS document.getElementById("listbox").offsetHeight is > 2
PASS successfullyParsed is true

TEST COMPLETE


Selecting menu list options using up/down key should fire onchange event

option 1
option 2
option 3
option 4
